pyproject.toml
[tool.poetry]
name = "fibsum"
version = "0.1.0"
description = ""
authors = ["amir <amirkhakshour@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-restful = "^0.3.8"
python-dotenv = "^0.14.0"
pytest = "^6.1.0"

[tool.poetry.dev-dependencies]
tox = "^3.20.0"
pytest = "^6.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
